% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actigraphr-package.R
\docType{package}
\name{actigraphr}
\alias{actigraphr}
\alias{actigraphr-package}
\title{actigraphr}
\description{
This package implements three standard algorithms for sleep detection from ActiGraph data: Sadeh, Cole-Kripke and Tudor-Locke.
}
\details{
In addition to the help pages, see the README page on \href{https://github.com/TheTS/actigraphr}{github} for examples.
}
