% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pa_troiano.R
\name{add_pa_troiano}
\alias{add_pa_troiano}
\title{Categorize activity into physical activity categories using Troiano cutpoints}
\usage{
add_pa_troiano(agdb, age = NA)
}
\arguments{
\item{agdb}{A \code{tibble} (\code{tbl}) of activity data (at least) an \code{epochlength} attribute.}

\item{age}{The age of the participant in whole years. This must be a value that is greater than or equal to 6 years. The function will first check the \code{agdb} for an appropriate age attribute, if it is not available, then an age must be explicitly specified.}
}
\value{
A \code{tibble} (\code{tbl}) of activity data. A new column \code{pa} of ordinal factor indicates the physical activity category for each epoch with labels "sed", "lig", "mod", "vig", and "ext".
}
\description{
Categorize activity into physical activity categories using Troiano cutpoints
}
\examples{
library("dplyr")
data("gtxplus1day")

gtxplus1day \%>\%
  add_pa_troiano()
}
