% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_choi.R
\name{apply_choi}
\alias{apply_choi}
\title{Apply the Choi algorithm}
\usage{
apply_choi(agdb, min_period_len = 90, min_window_len = 30,
  spike_tolerance = 2, use_magnitude = FALSE)
}
\arguments{
\item{agdb}{A \code{tibble} (\code{tbl}) of activity data (at least) an \code{epochlength} attribute. The epoch length must be 60 seconds.}

\item{min_period_len}{Minimum number of consecutive "zero" epochs to start a non-wear period. The default is 90.}

\item{min_window_len}{The minimum number of consecutive "zero" epochs immediately preceding and following a spike of artifactual movement. The default is 30.}

\item{spike_tolerance}{Also known as artifactual movement interval. At most \code{spike_tolerance} "nonzero" epochs can occur in sequence during a non-wear period without interrupting it. The default is 2.}

\item{use_magnitude}{Logical. If true, the magnitude of the vector (axis1, axis2, axis3) is used to measure activity; otherwise the axis1 value is used. The default is FALSE.}
}
\value{
A summary \code{tibble} of the detected non-wear periods. If the activity data is grouped, then non-wear periods are detected separately for each group.
}
\description{
The Choi algorithm detects periods of non-wear in activity data from an ActiGraph device. Such intervals are likely to represent invalid data and therefore should be excluded from downstream analysis.
}
\details{
The Choi algorithm extends the Troiano algorithm by requiring that short spikes of artifactual movement during a non-wear period are preceded and followed by \code{min_window_len} consecutive "zero" epochs.

This implementation of the algorithm expects that the epochs are 60 second long.
}
\examples{
file <- system.file("extdata", "GT3XPlus-RawData-Day01-10sec.agd",
                    package = "actigraph.sleepr")
agdb_10s <- read_agd(file)
agdb_60s <- collapse_epochs(agdb_10s, 60)
agdb_60s_nonwear <- apply_choi(agdb_60s)
}
\references{
L Choi, Z Liu, CE Matthews and MS Buchowski. Validation of accelerometer wear and nonwear time classification algorithm. \emph{Medicine & Science in Sports & Exercise}, 43(2):357–364, 2011.

ActiLife 6 User's Manual by the ActiGraph Software Department. 04/03/2012.
}
\seealso{
\code{\link{apply_troiano}}, \code{\link{collapse_epochs}}
}

