% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_cutpoints.R
\name{apply_cutpoints}
\alias{apply_cutpoints}
\title{Apply "cutpoints" to physical activity data}
\usage{
apply_cutpoints(agdb, cutpoints, use_magnitude = FALSE,
  custom_cutpoints = NULL)
}
\arguments{
\item{agdb}{A \code{tibble} of activity data optained from \code{\link{read_agd}}.}

\item{cutpoints}{A set of cutpoints to use. This can one of several predefined strings, or a custom list (see below).}

\item{use_magnitude}{Logical. If true, the vector magnitude is used to measure activity intensity; otherwise the axis1 value is used. The default is \code{FALSE}. Note that several predifined cutpoints will override the default (e.g. 'freedson_adult_vm').}

\item{custom_cutpoints}{A list of custom cutpoints. This must be a list containing two vectors. The first contains count threshold ranges (must be an even number), and the second contains a list of category names. These names are used as column names when using \code{\link{summarise_agd}}.See below for an example.}
}
\value{
The input \code{tibble} of activity data with an additional \emph{'activity'} column. The cutpoint information is appended to the tibble attributes.
}
\description{
Each epoch is assigned to an intensity category based on a set of "cutpoints".
}
\details{
All cutpoint values must be at the counts-per-minute (CPM) scale (60 sec epoch). However, you can still apply a set of cutpoints to data that has an \code{epochlength} of less that 60 seconds (the function adjusts the CPM values based on the \code{epochlength} attribute). The six sets of predefined cutpoints are:
\itemize{\item freedson_adult (1998)
         \item freedson_adult_vm (2011)
         \item freedson_children (2005)
         \item evenson_children (2008)
         \item mattocks_children (2007)
         \item puyau_children (2002)
         }
}
\examples{
data("gtxplus1day")

# Predefined cutpoints
agd <- gtxplus1day \%>\%
  apply_cutpoints("evenson_children")

# Custom cutpoints
my_cutpoints <- list(c(0, 100, 101, 1499, 1500, 3999, 4000, Inf),
                     c("sedentary", "light", "moderate", "vigorous"))

agd <- gtxplus1day \%>\%
  apply_cutpoints("custom",
                  custom_cutpoints = my_cutpoints)
}
\references{
Freedson, P. S., et al. (1998). Calibration of the Computer Science and Applications, Inc. accelerometer. \emph{Medicine and Science in Sports and Exercise} 30(5): 777-781.

Sasaki, J. E., et al. (2011). Validation and comparison of ActiGraph activity monitors. \emph{Journal of Science and Medicine in Sport} 14(5): 411-416.

Freedson, P., et al. (2005). Calibration of Accelerometer Output for Children. \emph{Medicine & Science in Sports & Exercise} 37(11): S523-S530.

Evenson, K. R., et al. (2008). Calibration of two objective measures of physical activity for children. \emph{Journal of sports sciences} 24(14): 1557-1565.

Mattocks, C., et al. (2007). Calibration of an accelerometer during free‐living activities in children. \emph{Pediatric Obesity} 2(4): 218-226.

Puyau, M. R., et al. (2002). Validation and calibration of physical activity monitors in children. \emph{Obesity} 10(3): 150-157.
}
