% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_sadeh.R
\name{apply_sadeh_}
\alias{apply_sadeh_}
\title{Internal function for calculating Sadeh algorithm sleep variables}
\usage{
apply_sadeh_(
  data,
  countname = "axis1",
  sleep_threshold = -4,
  adjustment = c("actilife", "pmax")
)
}
\arguments{
\item{data}{tibble of input data}

\item{countname}{character. Name of the vertical axis column}

\item{sleep_threshold}{numeric. Threshold above which the sleep index/sleep
probability will be classified as sleep}

\item{adjustment}{character. The method to use for avoiding \code{log(0)},
one of \code{"actilife"} (adds 1 count to all values) or \code{"pmax"} (adds
1 to all zero-count values and leaves others unchanged)}
}
\description{
Internal function for calculating Sadeh algorithm sleep variables
}
\keyword{internal}
