% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_weartime.R
\name{apply_weartime}
\alias{apply_weartime}
\title{Append a weartime column to an activity data frame}
\usage{
apply_weartime(agdb, fun = apply_troiano, ...)
}
\arguments{
\item{agdb}{A \code{tibble} of activity data obtained from \code{\link{read_agd}}.}

\item{fun}{The weartime function to employ (either \code{apply_troiano} or \code{apply_choi}). The default is \code{apply_troiano}.}

\item{...}{Additional parameters passed to the weartime function if the default parameters are not suitable.}
}
\value{
The input \code{tibble} of activity data with an additional \emph{'wear'} column.
}
\description{
This function wraps either the \code{\link{apply_choi}} or \code{\link{apply_troiano}} functions to calculate weartime, and appends a \emph{'wear'} column to the input \code{tibble}.
}
\examples{
 data("gtxplus1day")

 agd <- gtxplus1day \%>\%
  apply_weartime() \%>\%
  apply_cutpoints("evenson_children")

 # With additional arguments passed to the weartime function
 agd <- gtxplus1day \%>\%
  apply_weartime(fun = apply_troiano, min_period_len = 30) \%>\%
  apply_cutpoints("evenson_children")
}
\seealso{
\code{\link{apply_choi}}, \code{\link{apply_troiano}}
}
