% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_epochs_periods.R
\name{combine_epochs_periods}
\alias{combine_epochs_periods}
\title{Combine epochs with sleep/nonwear periods}
\usage{
combine_epochs_periods(epochs, periods, start_var, end_var)
}
\arguments{
\item{epochs}{A \code{tibble} (\code{tbl}) of activity data with (at least) a timestamp column.}

\item{periods}{A summary \code{tibble} of the (sleep or non-wear) periods.}

\item{start_var}{The \code{periods} column which specifies when the periods start.}

\item{end_var}{The \code{periods} column which specifies when the periods end.}
}
\value{
A \code{tibble} of activity data with one additional column, \code{period_id}, which indicates the period each epoch falls into.
}
\description{
Suppose we have used \code{apply_tudor_locke} to detect sleep periods or \code{apply_troiano}/\code{apply_choi} to detect non-wear periods. It might be useful to combine the epochs data with the periods data, so that each epoch is labeled according to which period it falls into, if any. Then we can easily slice the epochs data by sleep/non-sleep or wear/non-wear.
}
\examples{
library("dplyr")
data("gtxplus1day")

agdb <- gtxplus1day \%>\%
  collapse_epochs(60) \%>\%
  apply_sadeh()
periods <- agdb \%>\%
  apply_tudor_locke(min_sleep_period = 60)

agdb_with_periods <- combine_epochs_periods(agdb, periods,
                                            "in_bed_time", "out_bed_time")

# How many sleep periods were detected and what is their duration, in minutes?
periods \%>\% select(in_bed_time, out_bed_time, duration)
# What is the assignment of epochs to periods?
agdb_with_periods \%>\% count(period_id)
}
