% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complement_periods}
\alias{complement_periods}
\title{Find the complement of time periods}
\usage{
complement_periods(periods, epochs, start_var, end_var)
}
\arguments{
\item{periods}{A \code{tibble} with at least two columns,
\code{start_var} and \code{end_var} which are the first
and the last epoch in a set of time periods, e.g. sleep
periods or (non)wear periods.}

\item{epochs}{A \code{tibble} with at least one column,
\code{timestamp}, which contains POSIXct objects.}

\item{start_var}{The variable (unquoted) which indicates
when the time periods start.}

\item{end_var}{The variable (unquoted) which indicates when
the time periods end.}
}
\value{
A \code{tibble} of time periods with three columns:
\code{period_id} (a sequential identifier), \code{start_var}
(first epoch in period) and \code{end_var} (last epoch in period).
}
\description{
Find the complement of a set of time periods in a set of epochs.
For illustration, let's use integers instead of time periods and
epochs. Suppose we have two intervals/periods, \verb{\{[1,3], [8,10]\}};
their complement in the set \verb{\{[0, ..., 12]\}} is \verb{\{[0,0], [4,7], [11,12]\}}.
}
\examples{
library("lubridate")
library("dplyr")
periods <- tibble(
  start = ymd_hm("2017-01-01 00:01"),
  end = ymd_hm("2017-01-01 00:05")
)
epochs <- tibble(timestamp = ymd_hm("2017-01-01 00:00") +
  minutes(0:12))
complement_periods(periods, epochs, start, end)
}
