% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agd.R
\name{construct_period_polys}
\alias{construct_period_polys}
\title{Transform periods to polygons}
\usage{
construct_period_polys(periods, start_var, end_var, height)
}
\arguments{
\item{periods}{A \code{tibble} of periods with at least two columns \code{start_var} and \code{end_var}.}

\item{start_var}{The column which indicates when the time periods start.}

\item{end_var}{The column which indicates when the time periods end.}

\item{height}{The height of the rectangles.}
}
\description{
Transform periods to polygons, so that they can be overlaid on an activity plot.
}
