% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_timestamp}
\alias{expand_timestamp}
\title{Expand a time period into a vector of equally spaced time points}
\usage{
expand_timestamp(start, end, units = "1 min")
}
\arguments{
\item{start}{The start time, as a POSIXct object.}

\item{end}{The end time, as a POSIXct object.}

\item{units}{The time unit as a characters string.
The default is \code{"1 min"}.}
}
\value{
A POSIXct vector of equally spaced time points starting at
\code{start}, ending no later than \code{end}, and separated by \code{units}.
}
\description{
Given the start time and the end time of a period, expand it into a
vector of equally spaced time points.
}
\examples{
start <- as.POSIXct("2017-01-01")
end <- as.POSIXct("2017-01-01 01:00:00")
expand_timestamp(start, end, "15 mins")
}
