% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_epoch_length}
\alias{get_epoch_length}
\title{Guess the epoch length (in seconds) from the timestamp column}
\usage{
get_epoch_length(epochs)
}
\arguments{
\item{epochs}{A \code{tibble} with at least one column, \code{timestamp},
which contains POSIXct objects.}
}
\value{
The epoch length in seconds, inferred from the spacing between
consecutive timestamps. Throws an error if the time points are not
equally spaced.
}
\description{
Guess the epoch length (in seconds) from the timestamp column
}
\examples{
data("gtxplus1day")

gtxplus1day \%>\%
  get_epoch_length()

gtxplus1day \%>\%
  collapse_epochs(60) \%>\%
  get_epoch_length()
}
