% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtxplus1day}
\alias{gtxplus1day}
\title{GT3X+ sample data}
\format{
An object of class \code{tbl_agd} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8999 rows and 4 columns.
}
\usage{
gtxplus1day
}
\description{
24 hours of actigraphy data collected with a GT3X+ device,
which uses a tri-axial accelerometer to record activity in
vertical (axis 1), horizontal (axis 2) and lateral (axis 3) directions.
}
\keyword{datasets}
