% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_epochs.R
\name{has_missing_epochs}
\alias{has_missing_epochs}
\title{Checks whether there are gaps in the time series}
\usage{
has_missing_epochs(agdb)
}
\arguments{
\item{agdb}{A \code{tibble} (\code{tbl}) of activity data (at least)
an \code{epochlength} attribute.}
}
\value{
True or false.
}
\description{
The timestamps in the agd time series should run from
\code{first(timestamp)} to \code{last(timestamp)} in increments of
\code{epochlength} seconds. This function checks whether this holds
or not. If the data is grouped (e.g., by subject), the check is
performed for each group separately.
}
