% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agd.R
\name{plot_activity_summary}
\alias{plot_activity_summary}
\title{Plot activity summaries}
\usage{
plot_activity_summary(agdb_summary, start_date = NULL, end_date = NULL,
  colours = NULL, title = NULL)
}
\arguments{
\item{agdb_summary}{A \code{tibble} (\code{tbl}) of activity data obtained from  \code{\link{summarise_agd}}.}

\item{start_date}{The first day of data to plot. Default is \code{NULL} (plots all data).}

\item{end_date}{The last day of data to plot. efault is \code{NULL} (plots all data).}

\item{colours}{Colours are chosen by default, but a vector of custom colours can be passed in. The vector length should be equal to cutpoint categories + 1 (for non-wear time).}

\item{title}{A string used for the title of the plot. Defaults to the 'subjectname' attribute.}
}
\description{
Plot a grouped timeseries of activity values. The plot is organised (faceted) by date and coloured based on cutpoint categories.
}
\examples{
data("gtxplus1day")

agdb_summary <- gtxplus1day \%>\%
  apply_weartime() \%>\%
  apply_cutpoints("evenson_children") \%>\%
  summarise_agd(time = "1 hour")

plot_activity_summary(agdb_summary)

# Plot selected days only
plot_activity_summary(agdb_summary, start_date = '2012-06-27',
                            end_date = '2012-06-28')

# Plot with custom colours
plot_activity_summary(agdb_summary, colours = c(2:6))
}
