% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x.R
\name{read_acceleration}
\alias{read_acceleration}
\title{Read acceleration and timestamp from a log.bin file}
\usage{
read_acceleration(fid, accScale)
}
\arguments{
\item{fid}{A file connection able to read binary}

\item{accScale}{Scale factor to return acceleration in g}
}
\description{
Note, this passes a raw vector to c++ for parsing
}
