% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_activityC}
\alias{read_activityC}
\title{Convert a vector of raw bytes to yxz g values}
\usage{
read_activityC(bytes, scale)
}
\arguments{
\item{bytes}{A RawVector}

\item{scale}{Scale factor to return acceleration in g}
}
\value{
A NumericVector of g values in yxz order
}
\description{
Raw activity samples packed into 12-bit values in YXZ order.
}
