% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{read_agd_settings}
\alias{read_agd_settings}
\title{Read settings from an *.agd file}
\usage{
read_agd_settings(file, tz = "UTC")
}
\arguments{
\item{file}{Full path to an agd file to read.}

\item{tz}{Time zone to convert DateTime ticks to POSIX time.}
}
\description{
Read ActiGraph sleep watch data from a database stored in an AGD file. Return the settings.
}
\examples{
file <- system.file("extdata", "GT3XPlus-RawData-Day01.agd",
                    package = "actigraph.sleepr")
read_agd_settings(file)
}
