% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x.R
\name{read_info}
\alias{read_info}
\title{Read the info.txt file inside a .gt3x
This contains key items to help parse activity data}
\usage{
read_info(fid)
}
\arguments{
\item{fid}{A file connection to the info.txt}
}
\description{
Read the info.txt file inside a .gt3x
This contains key items to help parse activity data
}
