% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_agd.R
\name{summarise_agd}
\alias{summarise_agd}
\title{Summarises activity data over a specified time period}
\usage{
summarise_agd(agdb, time = "1 hour")
}
\arguments{
\item{agdb}{A \code{tibble} of activity data passed through \code{\link{read_agd}}, \code{\link{apply_weartime}}, and \code{\link{apply_cutpoints}}. This needs to contain an \emph{'activity'} and \emph{'wear'} column.}

\item{time}{The time interval that data is summarised. Valid options inclide \code{"sec", "min", "hour", "day"}. See \code{\link{cut.POSIXt}} for more details. The default time is \code{"1 hour"}.}
}
\value{
A \code{tibble} of summary activity data where rows correspond to the chosen \code{time} interval.
}
\description{
Summarises activity data over a time period.
}
\details{
The data are scaled to minutes (e.g. minutes of light intensity activity per hour) where appropriate. If the input \code{agdb} contains incline and steps data, this will also be summarised.
}
\examples{
 agd <- read_agd(file = "test.agd") \%>\%
  apply_weartime() \%>\%
  apply_cutpoints("freedson_children") \%>\%
  summarise_agd(time = "1 hour")

 # "min", "day" etc. are also valid time options
 summarise_agd(time = "30 min")
}
