% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_agd.R
\name{tbl_agd}
\alias{tbl_agd}
\title{A \code{tibble} of activity data exported by an ActiGraph device}
\usage{
tbl_agd(data, settings)
}
\arguments{
\item{data}{A \code{tibble} of raw activity counts.}

\item{settings}{A \code{tibble} of device settings.}
}
\value{
A \code{tibble} of activity data. It has a time index, \code{timestamp},
one or more activity value columns, and attributes holding settings
such as \code{epochlength}.
}
\description{
This tibble has several attributes, most importantly, \code{epochlength}.
}
\examples{
data("gtxplus1day")
gtxplus1day
attr(gtxplus1day, "epochlength")
}
