% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_period.R
\name{tbl_period}
\alias{tbl_period}
\title{A \code{tibble} (\code{tbl}) of automatically detected periods (of sleep or non-wear)}
\usage{
tbl_period(data)
}
\arguments{
\item{data}{A data frame of sleep or non-wear periods}
}
\description{
This tbl has several attributes which specify the detection algorithm and the its parameter settings.
}
