% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actigraph.sleepr-package.R
\docType{package}
\name{actigraph.sleepr}
\alias{actigraph.sleepr}
\alias{actigraph.sleepr-package}
\title{actigraph.sleepr}
\description{
This package implements three standard algorithms for sleep detection from ActiGraph data: Sadeh, Cole-Kripke and Tudor-Locke.
}
\details{
In addition to the help pages, see the README page on \href{https://github.com/dipetkov/actigraph.sleepr}{github} for examples.
}

