% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actigraph.sleepr-package.R
\docType{package}
\name{actigraph.sleepr}
\alias{actigraph.sleepr-package}
\alias{actigraph.sleepr}
\title{actigraph.sleepr}
\description{
This package implements three standard algorithms for sleep
detection from ActiGraph data: Sadeh, Cole-Kripke and Tudor-Locke.
}
\details{
In addition to the help pages, see the
\href{https://github.com/dipetkov/actigraph.sleepr}{README}
for examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dipetkov/actigraph.sleepr}
  \item Report bugs at \url{https://github.com/dipetkov/actigraph.sleepr/issues}
}

}
\author{
\strong{Maintainer}: Desislava Petkova \email{desislavka@gmail.com}

}
