% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_cole_kripke.R
\name{apply_cole_kripke}
\alias{apply_cole_kripke}
\title{Apply the Cole-Kripke algorithm}
\usage{
apply_cole_kripke(agdb)
}
\arguments{
\item{agdb}{A \code{tibble} (\code{tbl}) of activity data (at least) an
\code{epochlength} attribute. The epoch length must be 60 seconds.}
}
\value{
A \code{tibble} (\code{tbl}) of activity data. A new column
\code{sleep} indicates whether each 60s epoch is scored as asleep (S)
or awake (W).
}
\description{
The Cole-Kripke sleep scoring algorithm is primarily used for adult
populations as the supporting research was performed on subjects
ranging from 35 to 65 years of age.
}
\details{
The original paper proposes three versions of the Cole-Kripke algorithm,
optimized for 1-minute, 30-second and 10-second epochs. Here only the 1-min
version is implemented and therefore the \code{apply_cole_kripke} function
requires that the activity data is in 60s epochs. Use the
\code{\link{collapse_epochs}} function to modify higher-frequency data,
if necessary.

The Cole-Kripke algorithm uses the y-axis (axis 1) counts. First epoch
counts are divided by 100 and any scaled counts over 300 are clipped to 300.
This transformation is specific to ActiGraph devices. The sleep index (SI)
=======
\item{agdb}{A \code{tibble} (\code{tbl}) of activity data
(at least) an \code{epochlength} attribute. The epoch length must
be 60 seconds.}
}
\value{
A \code{tibble} (\code{tbl}) of activity data. A new
column \code{sleep} indicates whether each 60s epoch is scored
as asleep (S) or awake (W).
}
\description{
The Cole-Kripke sleep scoring algorithm is primarily used for
adult populations as the supporting research was performed on
subjects ranging from 35 to 65 years of age.
}
\details{
The Cole-Kripke algorithm requires that the activity data is in
60s epochs and uses a 7-minute window that includes the four
previous and two future epochs. This function implements the
algorithm as described in the ActiGraph user manual.

The Cole-Kripke algorithm uses the y-axis (axis 1) counts.
First epoch counts are divided by 100 and afterwards any
scaled counts over 300 are set to 300. The sleep index (SI)
>>>>>>> 545b00401120cb10b3280658026d795460f086b8
is defined as

\code{
.001 * (106 * epoch_prev(4) + 54 * epoch_prev(3) +
         58 * epoch_prev(2) + 76 * epoch_prev(1) +
        230 * epoch +
         74 * epoch_next(1) + 67 * epoch_next(2))
}

where at epoch \code{t}, \code{epoch_prev(i)} is the scaled activity count
\code{i} epochs before \code{t}. Similarly, \code{epoch_next(i)} is the
scaled activity count \code{i} epochs before \code{t}. That is, the algorithm
uses a 7-epoch window which includes the four preceding and the two
subsequent epochs. The time series of activity counts is padded with zeros as
necessary, at the beginning and at the end.

Finally, the sleep state is awake (W) if the sleep index SI is less than 1;
otherwise the sleep state is asleep (S).
}
\examples{
library("dplyr")
data("gtxplus1day")

gtxplus1day \%>\%
  collapse_epochs(60) \%>\%
  apply_cole_kripke()
}
\references{
RJ Cole, DF Kripke, W Gruen, DJ Mullaney and JC Gillin.
Automatic sleep/wake identification from wrist activity.
\emph{Sleep}, 15(5):461–469, 1992.

ActiLife 6 User's Manual by the ActiGraph Software
Department. 04/03/2012.
}
\seealso{
\code{\link{collapse_epochs}}, \code{\link{apply_sadeh}},
\code{\link{apply_tudor_locke}}
}
