% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_epochs.R
\name{impute_epochs}
\alias{impute_epochs}
\title{Impute missing count values}
\usage{
impute_epochs(agdb, ...)
}
\arguments{
\item{agdb}{A \code{tibble} of activity data with an \code{epochlength} attribute.}

\item{...}{Comma separated list of unquoted variables.}
}
\value{
A\code{tibble} of activity data. Each variable in \code{...} is imputed.
}
\description{
Trim leading and trailing NAs. Fill in the rest of the NAs using
cubic spline interpolation.
}
\examples{
library("dplyr")
data("gtxplus1day")

gtxplus1day$axis1[5:10] <- NA
gtxplus1day \%>\%
  impute_epochs(axis1)
}
\seealso{
\code{\link[zoo:na.approx]{zoo::na.spline()}}, \code{\link[zoo:na.trim]{zoo::na.trim()}}
}
