% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agd.R
\name{plot_activity}
\alias{plot_activity}
\title{Plot activity values}
\usage{
plot_activity(agdb, var, color = "black", nrow = NULL, ncol = NULL)
}
\arguments{
\item{agdb}{A \code{tibble} of activity data.}

\item{var}{The activity variable (unquoted) to plot on the y-axis.}

\item{color}{Activity line color.}

\item{nrow, ncol}{Number of rows and columns. Relevant only if the
activity data is grouped.}
}
\value{
A \code{ggplot} object. It shows the activity values in column \code{var}
as a time series, faceted by group when the input \code{agdb} is grouped.
}
\description{
Plot a time series of activity values (by default, the counts on the
vertical axis \emph{axis1}).
}
\examples{
data("gtxplus1day")
sub_gt3x <- gtxplus1day \%>\%
  dplyr::filter(timestamp <= lubridate::as_datetime("2012-06-27 18:00:00"))

data <- sub_gt3x \%>\%
  collapse_epochs(60) \%>\%
  apply_cole_kripke()

plot_activity(data, axis1, color = "gray")
plot_activity(data, axis1, color = "sleep")
}
