% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agd.R
\name{plot_activity}
\alias{plot_activity}
\title{Plot activity values}
\usage{
plot_activity(agdb, var = "axis1", color = "black", nrow = NULL,
  ncol = NULL)
}
\arguments{
\item{agdb}{A \code{tibble} (\code{tbl}) of activity data (at least) an \code{epochlength} attribute.}

\item{var}{The activity variable to plot. The default is axis1.}

\item{color}{Line color.}

\item{nrow, ncol}{Number of rows and columns. Relevant only if the activity data is grouped.}
}
\description{
Plot a time series of activity values (by default, the counts on the vertical axis \emph{axis1}).
}
\examples{
data("gtxplus1day")
plot_activity(gtxplus1day)
}
