% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agd.R
\name{plot_activity_period}
\alias{plot_activity_period}
\title{Plot activity and periods}
\usage{
plot_activity_period(
  agdb,
  periods,
  act_var,
  start_var,
  end_var,
  color = "black",
  fill = "#525252",
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{agdb}{A \code{tibble} of activity data.}

\item{periods}{A \code{tibble} of periods with at least two columns
\code{start_var} and \code{end_var}.}

\item{act_var}{The activity variable (unquoted) to plot on the y-axis.}

\item{start_var}{The variable (unquoted) which indicates when the
time periods start.}

\item{end_var}{The variable (unquoted) which indicates when the time
periods end.}

\item{color}{Activity line color.}

\item{fill}{Polygon fill color.}

\item{nrow, ncol}{Number of rows and columns. Relevant only if the
activity data is grouped.}
}
\value{
A \code{ggplot} object. It shows the activity values as a time series
and each period as a rectangle from \code{start_var} to \code{end_var}.
}
\description{
Plot activity values as a time series and periods as polygons.
}
\examples{
data("gtxplus1day")

sub_gt3x <- gtxplus1day \%>\%
  dplyr::filter(timestamp <= lubridate::as_datetime("2012-06-27 18:00:00"))
# Detect sleep periods using Sadeh as the sleep/awake algorithm
# and Tudor-Locke as the sleep period algorithm
periods_sleep <- sub_gt3x \%>\%
  collapse_epochs(60) \%>\%
  apply_cole_kripke() \%>\%
  apply_tudor_locke(min_sleep_period = 60)

plot_activity_period(
  sub_gt3x, periods_sleep, axis1,
  in_bed_time, out_bed_time
)
}
