% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{read_agd_raw}
\alias{read_agd_raw}
\title{Read an *.agd file, with no post-processing}
\usage{
read_agd_raw(file, tz = "UTC")
}
\arguments{
\item{file}{Full path to an agd file to read.}

\item{tz}{Time zone to convert DateTime ticks to POSIX time.}
}
\value{
A list of five tables: settings, data, filters, sleep and awakenings.
}
\description{
Read ActiGraph sleep watch data from an SQLite database stored in an AGD file. Return a list of five tables: data, sleep, filters, settings, awakenings.
}
\details{
These tables have the schema described in the ActiLife 6 User manual -- at least the column names, though not the data types. In particular, the timestamps are converted to human-readable POSIXct representation.

 In the SQLite database, timestamps are stored as ticks since 12:00:00 midnight, January 1, 0001. Each tick is one ten-millionth of a second and so ticks are of type INTEGER (long int). R does not have a 64 bit integer type and timestamps overflow. So I divide by 10,000,000 and convert to date/time with \code{STRFTIME}, \emph{before} selecting the these columns from the database.
}
\examples{
file <- system.file("extdata", "GT3XPlus-RawData-Day01-10sec.agd",
                    package = "actigraph.sleepr")
agdb <- read_agd_raw(file)
str(agdb)
}
\references{
ActiLife 6 User's Manual by the ActiGraph Software Department. 04/03/2012.

\code{covertagd}: R package for converting agd files from ActiGraph into data.frames.
}
\seealso{
\code{\link{read_agd}}
}

