% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_agd.R
\name{tbl_agd}
\alias{tbl_agd}
\title{A \code{tibble} (\code{tbl}) of activity data exported by an ActiGraph device}
\usage{
tbl_agd(data, settings)
}
\arguments{
\item{data}{A data frame of raw activity counts.}

\item{settings}{A data frame of device settings.}
}
\description{
This tbl has several attributes, most importantly - \code{epochlength}.
}

